requirements/base.txt
Django==4.0
django-environ==0.8.1
psycopg2-binary==2.9.2
djangorestframework==3.13.1
django-cors-headers==3.11.0
drf-spectacular
# requests
# weasyprint == 52.5